pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.3.0"
description = "Differences evaluator."
authors = ["Andrey Zhelezov <azhelezov@mail.ru>"]
packages = [
    { include = "gendiff"}
]

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff_script:main"

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
flake8 = "^6.0.0"
pytest = "^7.2.0"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
